### 2025-02-20 #7803 @editorialbot
[PRE REVIEW]: cuisto: A Python package to quantify histological data from QuPath and ABBA
**Submitting author:** @GuillaumeLeGoc (<a href="http://orcid.org/0000-0002-6946-1142">Guillaume Le Goc</a>)
**Repository:** https://github.com/TeamNCMC/cuisto
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 2025.1.28
