### 2025-02-21 #7803 @adamltyson
Hi @GuillaumeLeGoc,

I'll take over this submission from now on.

The first task is to find some (ideally three) reviewers. Would you be able to recommend appropriate reviewers for this submission? You can use the [list of potential reviewers](https://reviewers.joss.theoj.org/reviewers) to help. Please provide a list, but don't tag them (with @) so they aren't notified unnecessarily. It would be great to have one or two from outside the BrainGlobe world, I can find those myself :)

Let me know if you have any questions,
Adam
