### 2025-02-28 #7803 @GuillaumeLeGoc
Hi @Kevin-Mattheus-Moerman, thanks for starting the process ! I added some quick contributing guidelines. Let me know if that is enough.

Hello @adamltyson, thanks for accepting editing this work !
We identified 3 non-BrainGlobe potential reviewers that would be suitable from the list :
- [Antoine Ruzette](https://github.com/antoineruzette)
- [Yannan Chen](https://github.com/Yannan-Chen)
- [Pradeep Rajasekhar](https://github.com/pr4deepr)

Cheers,
Guillaume.
