### 2025-02-28 #7803 @adamltyson
Hi @antoineruzette, @Yannan-Chen, @igortatarnikov, @pr4deepr, @polarbean would you be able to review this submission to JOSS?

If you haven't reviewed for JOSS before, we carry out our checklist-driven reviews here in GitHub issues following [these guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). If you have any questions, let me know.

@pr4deepr, @Yannan-Chen, I know I've just asked you to review another submission, so no worries if you can't do both!
