### 2025-03-28 TeamNCMC/cuisto#17 @PolarBean
Please provide data for reproducability
https://github.com/openjournals/joss-reviews/issues/7906
Hi,
The example does not include paths to the files required for it to run. For instance here
https://github.com/TeamNCMC/cuisto/blob/main/examples/batch_process_animals.py#L24-L27
Can you please update the example to be self contained and runnable. If the data is too large to include in the github repository there are several options.
1. Include the data on an external data repository and fetch with a tool such as [pooch](https://github.com/fatiando/pooch)
2. Provide a paired down example which requries fewer data files.

Thanks
