### 2025-03-28 #7906 @PolarBean
@GuillaumeLeGoc The manuscript contains the following statement

> Regrettably however, these toolboxes often lack maintenance and/or community support, as is the case with the QUINT
> workflow (Yates et al., 2019) that relies on several pieces of software (including QuickNII
> (Puchades et al., 2019) and Ilastik (Berg et al., 2019)).

I am slightly biased as I am a developer on the QUINT workflow but what is meant by unmaintained? And how is lack of community support quantified?

Thank you in advance.
