### 2025-03-28 TeamNCMC/cuisto#18 @PolarBean
Highlight usecases
https://github.com/openjournals/joss-reviews/issues/7906
Consider updating either the documentation or readme to highlight how this tool is useful. I had to dig into the example notebooks to understand how the tool could be used. Consider lifting some of the quantification tables and graphs either into the readme or first page of documentation.

The Quickstart guide is too focused on installation and should instead also include a basic example. The long installation guide should be under a seperate tab. For reference, a related tool which achieves this nicely is [brainrender](https://github.com/brainglobe/brainrender) which immediately includes a large figure conveying to the user potential uses of the tool.
