### 2025-03-28 TeamNCMC/cuisto#19 @PolarBean
Getting started should show more than installation
https://github.com/openjournals/joss-reviews/issues/7906
On the getting started page it would be nice to quickly show how the tool could be used. For inspiration see
https://pandas.pydata.org/docs/user_guide/10min.html
and
https://numpy.org/doc/stable/user/quickstart.html

The more detailed installation explanation should be tucked away in a seperate tab.
