### 2025-03-28 #7906 @PolarBean
@GuillaumeLeGoc The manuscript mentions the BriaAn tool and that it has some overlap with cuisto, however it's not clear to me how significant this overlap is. Could you please expand on this in the manuscript, ideally describing in which situations one would choose cuisto over BriaAn.
