### 2025-03-28 #7906 @PolarBean
@GuillaumeLeGoc On line 71 the manuscript states "the areal density" I assume this is a typo and is intended to say "the area density"
