### 2025-03-28 #7906 @PolarBean
@GuillaumeLeGoc Upon delving into the repository and manuscript in more detail it appears to me that what you are providing are
Instructions on how to install and use ABBA and QuPath, and scripts for plotting the results. What is unclear to me is how much processing is performed by cuisto? Is cuisto performing more than plotting? If so this is not made clear, could you please clearly highlight all the functionality provided by cuisto?
