### 2025-03-28 #7906 @PolarBean
@GuillaumeLeGoc  Regarding the animal ethics check list, the reposiitory contains multiple files which I assume are derived from mouse brains. Could you please add a statement to the manuscript regarding the animal ethics approval for this. If this data is derived from a seperately published study which has ethics approval please indicate this so that I may check this off. Thank you.
