### 2025-04-14 #7906 @GuillaumeLeGoc
Hi @adamltyson , @PolarBean  and @IgorTatarnikov , sorry for the delay getting back to you.

Many thanks for taking the time to review our work.
Here, I will address the points raised by @PolarBean here and the [repository](https://github.com/TeamNCMC/cuisto/issues), and write down what could be modified to improve this submission. @IgorTatarnikov, I don't know if you plan to comment further on those issues nor if you identified additionnal points -- we're all ears.

### Comment 1
> [@GuillaumeLeGoc](https://github.com/GuillaumeLeGoc) The manuscript contains the following statement
>
> > Regrettably however, these toolboxes often lack maintenance and/or community support, as is the case with the QUINT
> > workflow (Yates et al., 2019) that relies on several pieces of software (including QuickNII
> > (Puchades et al., 2019) and Ilastik (Berg et al., 2019)).
>
> I am slightly biased as I am a developer on the QUINT workflow but what is meant by unmaintained? And how is lack of community support quantified?
>
> Thank you in advance.

I acknowledge that the initial wording may be a bit harsh and apologize as it was not our intent. For the context, when I started to work on 2D brain-wide mapping of neuronal processes, I reviewed existing software solutions for this task and two stood out: ABBA and QUINT. But delving into the latter, I was confused by the following points, which drove us in part to use ABBA instead.
Most limits relate to an unclear governance and no centralized documentation and access to files :
- QuickNII (and VisuAlign) exists in the [HumanBrainProject](https://github.com/HumanBrainProject/QuickNII) org, the [Neural-Systems-at-UIO](https://github.com/Neural-Systems-at-UIO/QuickNII) org and in a [user repo](https://github.com/Tevemadar/QuickNII)
- The QUINT documentation contains a tip to get more information from the [EBRAINS](https://www.ebrains.eu/brain-atlases/analysis/labelled-features-analysis/) website which in turns has a "Learn more" link back to the documentation
- Links to the GitHub repositories in the documentation points to the Neural-Systems-at-UIO organization, and links from EBRAINS go to HumanBrainProject repositories
- Downloads happen over at [NITRC website](https://www.nitrc.org/projects/quicknii/)

Furthermore, timestamps do not suggest an active maintenance:
- QuickNII and VisuAlign repositories did not get updates since 6 years (not counting docs and readme files)
- [VisuAlign documentation](https://visualign.readthedocs.io/en/latest/index.html) is quite thin, and contains a "Coming soon" mention on a 2021-copyrighted page.

Overall, this un unclear organization leads to difficulties to know where to get support, and making it difficult to quantify the lack of community support. Image.sc forum is not mentioned in the QUINT documentation, but there, as of 2025-04-14, #quicknii and #quint tags cumulate 21 mentions while #abba gets 214.

- [ ] We'd be happy to rephrase and/or include details.

### Comment 2
> [@GuillaumeLeGoc](https://github.com/GuillaumeLeGoc) The manuscript mentions the BriaAn tool and that it has some overlap with cuisto, however it's not clear to me how significant this overlap is. Could you please expand on this in the manuscript, ideally describing in which situations one would choose cuisto over BriaAn.

Unfortunately, the timing did not allow us to review BraiAn and make a proper comparison between the two tools. I would say "the more, the merrier" as in my opinion having multiple software choices is good for the end user. As the main contributor of `cuisto`, I can only be (very) biased but looking at the documentation of BraiAn it seems a little more complicated to configure, especially on the QuPath side.
Therefore, `cuisto`, by its design, might be easier to learn and use for quick data pooling and plotting from custom QuPath analysis while still giving the opportunity to further analyse the data saved by `cuisto`. On the other hand BraiAn might be more adapted if one wants a full pipeline right from the start with configurable QuPath projects management.

- [ ] We can expand a bit on BraiAn but as it is still in development and in review I would not go ahead and make a comparative study.

### Comment 3
> [@GuillaumeLeGoc](https://github.com/GuillaumeLeGoc) Upon delving into the repository and manuscript in more detail it appears to me that what you are providing are Instructions on how to install and use ABBA and QuPath, and scripts for plotting the results. What is unclear to me is how much processing is performed by cuisto? Is cuisto performing more than plotting? If so this is not made clear, could you please clearly highlight all the functionality provided by cuisto?

Related (https://github.com/TeamNCMC/cuisto/issues/19) :
> [openjournals/joss-reviews#7906](https://github.com/openjournals/joss-reviews/issues/7906) On the getting started page it would be nice to quickly show how the tool could be used. For inspiration see https://pandas.pydata.org/docs/user_guide/10min.html and https://numpy.org/doc/stable/user/quickstart.html
>
> The more detailed installation explanation should be tucked away in a seperate tab.

Related (https://github.com/TeamNCMC/cuisto/issues/18) :
> [openjournals/joss-reviews#7906](https://github.com/openjournals/joss-reviews/issues/7906) Consider updating either the documentation or readme to highlight how this tool is useful. I had to dig into the example notebooks to understand how the tool could be used. Consider lifting some of the quantification tables and graphs either into the readme or first page of documentation.
>
> The Quickstart guide is too focused on installation and should instead also include a basic example. The long installation guide should be under a seperate tab. For reference, a related tool which achieves this nicely is [brainrender](https://github.com/brainglobe/brainrender) which immediately includes a large figure conveying to the user potential uses of the tool.

We agree that at this time it might not be straightforward to understand what `cuisto` does and what one could expect from the tool.

Before plotting, `cuisto` pools the data (brain regions appear on several slices and need to be pooled together, summing their area), derives quantifying metrics (raw count or cumulated length, density per surface or relative density) as well as spatial distributions, gets the subjects averages and errors and plots the results.
This is highlighted on the 4 bullet points on the landing page of the documentation. Furthermore, we tried to highlight the role of `cuisto` and how it interacts with ABBA and QuPath on the color-coded block diagram on this same page.
The scripts you are referring to are examples that uses `cuisto`, which is a package with a documented API. Finally, the documentation showcases examples in the form Jupyter notebooks with included toy data to illustrate various use-cases.

When drafting the documentation for `cuisto`, it felt weird to only include the latter as it heavily relies on QuPath and ABBA. We hence decided to also include instructions for those softwares, and on how to format the data to easily use `cuisto` afterwards.

We will attempt to make this clearer by :

- [ ] Including an example graph on the landing page.
- [ ] Moving the ["Slow start"](https://teamncmc.github.io/cuisto/main-getting-started.html#slow-start) section on a dedicated page.
- [ ] Expanding the [Quick start](https://teamncmc.github.io/cuisto/main-getting-started.html#quick-start) section with bits from the [Using cuisto](https://teamncmc.github.io/cuisto/guide-use-cuisto.html#using-cuisto) page and [examples](https://teamncmc.github.io/cuisto/demo_notebooks/fibers_length_multi.html).

### Comment 4
> [@GuillaumeLeGoc](https://github.com/GuillaumeLeGoc) Regarding the animal ethics check list, the repository contains multiple files which I assume are derived from mouse brains. Could you please add a statement to the manuscript regarding the animal ethics approval for this. If this data is derived from a seperately published study which has ethics approval please indicate this so that I may check this off. Thank you.

Thanks for raising this, this was indeed overlooked.
- [ ] An animal ethics statement will be included in the manuscript.

### External comment
https://github.com/TeamNCMC/cuisto/issues/17
> [openjournals/joss-reviews#7906](https://github.com/openjournals/joss-reviews/issues/7906) Hi, The example does not include paths to the files required for it to run. For instance here https://github.com/TeamNCMC/cuisto/blob/main/examples/batch_process_animals.py#L24-L27 Can you please update the example to be self contained and runnable. If the data is too large to include in the github repository there are several options.
>
>     1. Include the data on an external data repository and fetch with a tool such as [pooch](https://github.com/fatiando/pooch)
>
>     2. Provide a paired down example which requries fewer data files.
>
>
> Thanks

Data for example notebooks are included in the [resources](https://github.com/TeamNCMC/cuisto/tree/main/resources) directory of the repository. Notebooks are [here](https://github.com/TeamNCMC/cuisto/tree/main/docs/demo_notebooks) and are self-runnable (they are run when the doc is built).

The example script you are referring to is here for the use case described [here](https://teamncmc.github.io/cuisto/guide-use-cuisto.html#examples).

- [ ] We will add a tsv file in the resources directory to mimick a multi-animal, non-pipeline situation and use those files in the script.
