### 2025-04-15 #7906 @PolarBean
Thank you for your thoughtful reply.

> Before plotting, cuisto pools the data (brain regions appear on several slices and need to be pooled together, summing their area), derives quantifying metrics (raw count or cumulated length, density per surface or relative density) as well as spatial distributions, gets the subjects averages and errors and plots the results.

Thank you for this clarification. Given that the tool performs collating of data across regions and plotting this without additional spatial analysis (for instance translating the points into 3D coordinates seems to be handled by other tools) I am not sure that this is substantial enough to warrant publication at present. Though I will leave this decision to @adamltyson
