### 2025-04-15 #7906 @adamltyson
Hi @PolarBean, the criteria for whether something warrants pubication in JOSS is what we call "substantial scholarly effort" and is detailed here: https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort.

Would you mind taking a look at this and confirming that you don't think this submission meets these criteria?

@IgorTatarnikov along with your review, would you mind also looking at these criteria and letting me know what you think?

Thanks all.
