### 2025-04-15 #7906 @PolarBean
`Whether the software is sufficiently useful that it is likely to be cited by your peer group.`
I would perhaps question the utility of collation and plotting alone and would likely classify this as a minor utility package. I am open to changing my mind if @IgorTatarnikov disagrees.
