### 2025-04-28 #7906 @adamltyson
Hi @GuillaumeLeGoc, did you have any thoughts about @IgorTatarnikov's suggestion. On face value it seems sensible, but I haven't dug into the software myself.
