### 2025-04-28 #7906 @GuillaumeLeGoc
Hi @adamltyson, I was waiting for Igor to finish his review before commenting further.

There are a bunch of scripts in the `scripts` folder, and I don't know which one(s) Igor is referring to. Most of them are simple QuPath utility scripts, including those to format data for `cuisto`. The Python segmentation one is a wrapper around methods available in the `cuisto` package. The brain mask detection scripts are basic image processing steps to prepare images for QuPath and ABBA.

In any case, full integration with QuPath is out of scope -- and [BraiAn](https://silvalab.codeberg.page/BraiAn/) provides such a tool. We aimed at a simpler process that integrates in researchers wokflow, allowing for accessible data collation, metrics *computation* and display with minimal programming skill -- as the latter is not always readily available in a team; and providing a reference documentation on how to proceed along the way. This aim, besides accessibility, also allows the user to manually interact with the data during each steps. While I agree that an automated, headless, batch process would be ideal, it is often -- if not always -- out of reach due to the variability observed in the imaging data (that can be due to animal , viral vector effectiveness, staining, cutting, ... variability).

While we think there might be an interest from the community for the functionnality provided by the utility scripts, the documentation and the `cuisto` package, as evidenced by questions from users asking for help over a the [image.sc forum](https://image.sc) to both find objects with QuPath, script some processes and extract per-regions metrics from the CSV tables QuPath can output, if this submission is deemed to not be substantial enough, let us know.

Cheers,
Guillaume
