### 2025-05-05 TeamNCMC/cuisto#19 @IgorTatarnikov
https://github.com/openjournals/joss-reviews/issues/7906

I agree with @PolarBean here, I would also suggesting adding an example output (say the graphs from the `cells_distributions` notebook) to the Introduction page.
