### 2025-05-05 TeamNCMC/cuisto#20 @IgorTatarnikov
Standardise how scripts are linked in the docs
https://github.com/openjournals/joss-reviews/issues/7906

References to the scripts should contain the full path to the script in question in the documentation. e.g. scripts/qupath-utils/pipelineImportExport.groovy instead of piplineImportExport.groovy. It would also be nice to link to the script whenever it's mentioned to make navigation easy.

I also found one link for `segment_image.py` that points to an API doc rather than the GItHub repo, is this intentional? If so, then other scripts should also be documented in the same manner for consistency.
