### 2025-05-05 #7906 @IgorTatarnikov
I've finished with my review, apologies for the delay.

Overall the project is of good quality, and I would suggest accepting with minor revisions.

### Documentation feedback
The documentation is of good standard overall, however, I agree with @PolarBean that the Introduction and Getting Started pages would benefit from providing example outputs, and use cases for the package as one of the first points (https://github.com/TeamNCMC/cuisto/issues/19).

Further, since scripts are external to the package but heavily featured in the documentation I would suggest standardising how they're linked throughout https://github.com/TeamNCMC/cuisto/issues/20.

### Functionality feedback

It's good to see an automated integration test for `process_animal` and `process_animals`, however I would suggest adding unit tests to test functionality in a more granular manner. This is optional but highly suggested.

> There are a bunch of scripts in the `scripts` folder, and I don't know which one(s) Igor is referring to. Most of them are simple QuPath utility scripts, including those to format data for `cuisto`. The Python segmentation one is a wrapper around methods available in the `cuisto` package. The brain mask detection scripts are basic image processing steps to prepare images for QuPath and ABBA.

I was mostly referring to the `generate_atlas_outlines.py` and `segment_images.py`.

1) `generate_atlas_outlines.py`
From my understanding plotting 2D heatmaps is one of the core functionalities of `cuisto` and it felt disjointed to download a script in order to generate the required `outlines.h5` file. If the 2D heatmaps are core, then the functionality in `generate_atlas_outlines` should be included and shipped in the main package. Further, would it be possible to pre-generate the outline files and host them with [Git LFS](https://docs.github.com/en/repositories/working-with-files/managing-large-files/about-git-large-file-storage) or [GIN](https://gin.g-node.org/)? The outlines can be fetched and stored in a hidden directory when the user first requests them for a given atlas.

2) `segment_images.py`
I recognise that the script is mostly a wrapper around other methods in the `cuisto` pacakage, however, I think `process_directory` and the corresponding helper functions should be available to users of the core package without the need to download the `segment_images.py` script.

Best,
Igor
