### 2025-05-05 TeamNCMC/cuisto#21 @IgorTatarnikov
Add unit tests
https://github.com/openjournals/joss-reviews/issues/7906

I would suggest adding unit tests for functions in the package to increase the granularity of coverage and testing feedback.
