### 2025-05-05 TeamNCMC/cuisto#22 @IgorTatarnikov
Move 2D atlas outline creation into the core package
https://github.com/openjournals/joss-reviews/issues/7906

Currently, users are unable to plot 2D heatmaps until they download and execute the `generate_atlas_outlines.py` script. This functionality should be moved into the package. I would also suggest pre-generating the atlas outlines and storeing them in [Git LFS](https://docs.github.com/en/repositories/working-with-files/managing-large-files/about-git-large-file-storage) or [GIN](https://gin.g-node.org/) to fetch when requested by the user.
