### 2025-05-05 TeamNCMC/cuisto#23 @IgorTatarnikov
Move batch processing functionality from `segment_images.py` script into the main package
https://github.com/openjournals/joss-reviews/issues/7906

I think the batch processing functionality defined in the `segment_images.py` script should be moved into the main package. As segmenting probability maps is part of the `cuisto` package, users who install `cuisto` from `pip` or `conda-forge` should have access to the batch processing capabilites.
