### 2025-05-05 #7906 @PolarBean
Also one thing I would suggest is given part of the contribution here is improved installation instructions for other packages, these would be more suitable as pull requests to the packages in question in order to improve their documentation.
