### 2025-05-13 TeamNCMC/cuisto#20 @GuillaumeLeGoc
Thanks for the suggestion. We actually recommended that in the [contributing guidelines](https://github.com/TeamNCMC/cuisto/blob/main/CONTRIBUTING.md#documentation) -- which it seems I didn't always follow.
Usually, unless overlooked, the Groovy scripts should be mentionned like this :

> In [`scripts/qupath-utils/segmentation`](https://github.com/TeamNCMC/cuisto/tree/main/scripts/qupath-utils/segmentation), there is `watershedDetectionFilters.groovy` [...]

Where the location has a backlink.
The [Pipeline page](https://teamncmc.github.io/cuisto/guide-pipeline.html) do lack backlinks, which will be fixed shortly.
