### 2025-05-14 9e2c35a
Documentation update * modified all references to the old segment_images.py script to the new segmentation module * mention the existence of an example usage script in the examples folder * some typos here and there
