+++ docs/guide-qupath-objects.md
@@ -132,7 +132,7 @@ The `Create objects` action will ask what where the objects should be created. I
     In [`scripts/qupath-utils/segmentation`](https://github.com/TeamNCMC/cuisto/tree/main/scripts/qupath-utils/segmentation), there is a `createDetectionsFromPixelClassifier.groovy` script to batch-process your project.
 
 #### Probability map segmentation
-Alternatively, a Python script provided with `cuisto` can be used to segment the probability map generated by the pixel classifier (the script is located in `scripts/segmentation`).
+Alternatively, the [`cuisto.segmentation`](api-segmentation.md) module can be used to segment the probability map generated by the pixel classifier (there is an example usage [here](https://github.com/TeamNCMC/cuisto/blob/main/examples/segment_images.py)).
 
 You will first need to export those with the `exportPixelClassifierProbabilities.groovy` script (located in [`scripts/qupath-utils/segmentation`](https://github.com/TeamNCMC/cuisto/tree/main/scripts/qupath-utils/segmentation)).
 
@@ -141,7 +141,7 @@ Then the segmentation script can :
 + find punctual objects as polygons (with a shape) or points (punctual) that can be counted.
 + trace fibers with skeletonization to create lines whose lengths can be measured.
 
-Several parameters have to be specified by the user,  see the segmentation script [API reference](api-script-segment.md). This script will generate [GeoJson](tips-formats.md#json-and-geojson-files) files that can be imported back to QuPath with the [`importGeojsonFiles.groovy`](https://github.com/TeamNCMC/cuisto/blob/main/scripts/qupath-utils/segmentation/importGeojsonFiles.groovy) script.
+Several parameters have to be specified by the user, see the segmentation script [API reference](api-script-segment.md). This script will generate [GeoJson](tips-formats.md#json-and-geojson-files) files that can be imported back to QuPath with the [`importGeojsonFiles.groovy`](https://github.com/TeamNCMC/cuisto/blob/main/scripts/qupath-utils/segmentation/importGeojsonFiles.groovy) script.
 
 #### Other use of the pixel classifier
 As you might have noticed, when loading a pixel classifier in your project, 3 actions are available. "Create objects" is described [above](#built-in-create-objects), which leaves the other two.
