+++ docs/api-segmentation.md
@@ -1,0 +1,1 @@
+:::cuisto.segmentation
