### 2025-05-15 TeamNCMC/cuisto#22 @GuillaumeLeGoc
This would be nice to have indeed.
Outlines for `allen_mouse_10um` and `allen_cord_20um` were pre-generated and are hosted in our facility, the links are in the [README](https://github.com/TeamNCMC/cuisto?tab=readme-ov-file#brain-structures). Those links will be referenced in the [Brain contours](https://teamncmc.github.io/cuisto/tips-brain-contours.html) section of the documentation.
Unfortunately, generating those for all available Brainglobe atlases seems a bit too intensive. I will include the outline generation in the package, and they could be generated on the fly when needed - although this takes a while a regular computer.
I will explore Git LFS or GIN so we may host some standard outlines (eg. `allen_mouse_10um`).
