### 2025-05-15 TeamNCMC/cuisto#22 @GuillaumeLeGoc
Update : I tried to implement the outline generation on the fly before actually plotting, but I didn't notice how much RAM it requires - it crashes on my 16GB laptop. So I'll just print a message stating why the contours will not be drawn if the file is not specified in the configuration file.
