+++ cuisto/__init__.py
@@ -2,7 +2,7 @@
 Perform quantification of objects in registered and segmented histological slices.
 """
 
-from . import compute, display, io, process, utils, seg
+from . import atlas, compute, display, io, process, utils, seg
 from .config import Config
 
-__all__ = ["Config", "compute", "display", "io", "process", "utils", "seg"]
+__all__ = ["Config", "atlas", "compute", "display", "io", "process", "utils", "seg"]
