### 2025-05-15 1f24432
Include brain structures outline generation + added the atlas module + removed the generate_atlas_outlines.py script + added info message if the outline file is not found while plotting + allow disabling outline plotting if no structures are specified
