### 2025-05-16 130dc0f
Download outlines if available + attempt to download outlines files + added explicit requests dep + updated doc to reflect the new behavior
