+++ configs/config_template.toml
@@ -82,5 +82,5 @@ log_scale = false  # use log. scale for metrics
 [files]  # full path to information TOML files
 blacklist = "../../atlas/atlas_blacklist.toml"
 fusion = "../../atlas/atlas_fusion.toml"
-outlines = "/data/atlases/allen_mouse_10um_outlines.h5"
+outlines = ""  # download and change the path or leave empty if the outlines are available to download
 infos = "../../configs/infos_template.toml"
