### 2025-05-16 TeamNCMC/cuisto#22 @GuillaumeLeGoc
Should be addressed with #26
