### 2025-05-18 cb6e21a
Documentation improvements + moved installation slow start section in a dedicated guide + added examples in the quick start page
