### 2025-05-18 44d4fc5
Update example to batch process files exported individually from QuPath to get example data online
