+++ examples/batch_process_animals.py
@@ -1,7 +1,7 @@
 """
 This example shows how to collect data from single animal when they were exported
 individually from QuPath, resulting in a single file per animal, containing data from
-all the image in the each project. Thus, there should be, for each animal, a file
+all the images in each project. Thus, there should be, for each animal, a file
 corresponding to Annotations (brain regions) and Detections (objects of interest).
 
 We assume all the pairs of files are located in the same directory, and their file name
@@ -16,17 +16,57 @@ script. We assume all json files corresponding to one animal is stored in a
 
 # import required packages
 import os
+import tarfile
+import sys
+from pathlib import Path
 
-import cuisto
 import pandas as pd
+import requests
 from tqdm import tqdm
 
+import cuisto
+
 # --- Parameters
-input_dir = "/path/to/tsv/files"
-animals = ("animalid0", "animalid1", "animalid2")
-config_file = "/path/to/config/file.toml"
+# Configure to get the example online
+dl_example = True
+animals = ("animalid0", "animalid1")
+example_url = "https://github.com/TeamNCMC/cuisto/raw/main/resources/example.tar.gz"
+
+# Configure with your own data
+# full path to where are the TSV files and {animal}_detections folders
+# dl_example = False
+# animals = ("animalid0", "animalid1")
+# input_dir = /path/to/your/data
+# config_file = /path/to/your/config.toml
+
+# Download example data into cuisto default folder. Remove this block if using your data
+if dl_example:
+    default_destination = Path.home() / ".cuisto"
+    print(f"Downloading example data to {default_destination}...")
+    if not default_destination.exists():
+        default_destination.mkdir()
+
+    response = requests.get(example_url)
+    if response.ok:
+        tarname = default_destination / "example.tar.gz"
+        with open(tarname) as fid:
+            fid.write(response.content)
+        with tarfile.open(tarname) as tar:
+            tar.extractall(tarname.parent, filter="data")
+    else:
+        msg = (
+            "Download failed. Download manually here : "
+            "https://github.com/TeamNCMC/cuisto/tree/main/resources"
+        )
+        print(msg)
+        sys.exit()
+
+    # set paths
+    input_dir = default_destination / "example"
+    config_file = input_dir / "config_multi.toml"
 
 # --- Preparation
+
 # load configuration
 cfg = cuisto.Config(config_file)
 
@@ -49,18 +89,18 @@ for animal in pbar:
     )
     # read detections only to plot spatial distributions, otherwise set
     # df_detections = pd.DataFrame()
-    # comment out for fibers
-    df_detections = pd.read_csv(
-        os.path.join(input_dir, f"{animal}_detections.tsv"),
-        index_col="Object ID",
-        sep="\t",
-    )
-    # uncomment for fibers
-    # df_detections = cuisto.io.cat_json_dir(
-    #     os.path.join(input_dir, f"{animal}_detections"),
-    #     hemisphere_names=cfg.hemispheres["names"],  # we need it now for performance
-    #     atlas=cfg.bg_atlas,
+    # uncomment out for cells
+    # df_detections = pd.read_csv(
+    #     os.path.join(input_dir, f"{animal}_detections.tsv"),
+    #     index_col="Object ID",
+    #     sep="\t",
     # )
+    # comment out for non-fibers
+    df_detections = cuisto.io.cat_json_dir(
+        os.path.join(input_dir, f"{animal}_detections"),
+        hemisphere_names=cfg.hemispheres["names"],  # we need it now for performance
+        atlas=cfg.bg_atlas,
+    )
 
     # get results
     df_reg, dfs_dis, df_coo = cuisto.process.process_animal(
