### 2025-05-18 9759028
Update batch_process_animals.py + fixed writing downloaded data + fixed wrong path in config.toml in the example tar
