+++ examples/batch_process_animals.py
@@ -49,7 +49,7 @@ if dl_example:
     response = requests.get(example_url)
     if response.ok:
         tarname = default_destination / "example.tar.gz"
-        with open(tarname) as fid:
+        with open(tarname, "wb") as fid:
             fid.write(response.content)
         with tarfile.open(tarname) as tar:
             tar.extractall(tarname.parent, filter="data")
