### 2025-05-18 1ea1794
Update batch_process_animals.py + fixed configuration files path in the downloaded configuration file...
