### 2025-05-18 TeamNCMC/cuisto#17 @GuillaumeLeGoc
Hi,
This is done, the new example can be set up to download a minimal archive with example data (hosted in the resources directory), configuration files and so on, that mimicks the use-case where the user export manually individual animal in QuPath and wants to average them.

(I don't reference a PR because I forgot to branch before editing...)
