### 2025-05-18 #7906 @PolarBean
Hey @GuillaumeLeGoc Thanks for making those changes it's much better. I understand if this is a conflict of interest since I work on these tools @adamltyson however I think some of these points in the manuscript about QUINT are factually wrong.

` QuickNII and VisuAlign repositories did not get updates since 6 years (not counting docs and readme files)`
The last non-docs / readme commit to visualign was on March of this year. https://github.com/Neural-Systems-at-UIO/VisuAlign/commit/fee5f3d90fc53f15461aae7682fc041a9df77d1f

`Overall, this un unclear organization leads to difficulties to know where to get support, and making it difficult to quantify the lack of community support. Image.sc forum is not mentioned in the QUINT documentation, but there, as of 2025-04-14, #quicknii and #quint tags cumulate 21 mentions while #abba gets 214.`
QUINT doesnt use Image.sc but rather github in addition to the EBRAINS support system. Maybe number of publications succesfully using the tool might be a better metric but I think this is a rabbit hole and for a different paper.

I also think your tool is performing a different task to QUINT (you are plotting whereas quint is registration) and so the comparison is confusing.

Overall with your  changes cuisto is much improved. I will go through and test this week.

Harry
