### 2025-05-19 #7906 @GuillaumeLeGoc
Hi @PolarBean , thanks for your comment.

In my humble opinion I don't see this as a conflict of interest.

Indeed you're right, I missed the updates on VisuAlign (on the Neural-Systems-at-UIO repo, the one from EBRAINS lags behind). I mainly used the QuickNII repo as a reference after digging into existing solutions back in May 2023. I did try to look at publications but since ABBA does not have one (yet) it is hard to compare, and I confess I quickly abandoned as it is hard to track the QUINT workflow as a whole (since there are several corresponding papers).

The idea we wanted to convey is that a pipeline for registration + object detections + collating/computing/plotting does already exist (QuickNII/VisuAlign/DeepSlice + Ilastik + Nutil), but using QuPath and ABBA instead does not provide the last item, and `cuisto` aims at filling this gap. We are happy to rephrase to make this clearer and remove the indeed false statement about QUINT maintenance status.

Cheers,
Guillaume
