### 2025-05-19 #7906 @adamltyson
Sounds like there's agreement here. I would remove the image.sc comparison too, as it would only be fair if all tools had officially adopted image.sc as the primary support platform.
