### 2025-05-19 #7906 @PolarBean
Hey @adamltyson I'll go through and test the changes @GuillaumeLeGoc made in response to our issues this week. If all works as expected I'll complete the checklist.
