### 2025-05-22 TeamNCMC/cuisto#30 @PolarBean
excessive dependencies
https://github.com/openjournals/joss-reviews/issues/7906
Hi just retesting everything and noticed that there are 179 dependencies that are installed along with this package. Is it possible to trim this down somewhat? Feel free to say no if you feel this is unreasonable.
