### 2025-05-22 TeamNCMC/cuisto#31 @PolarBean
example produces unclear plots
github.com/openjournals/joss-reviews/issues/7906
![Image](https://github.com/user-attachments/assets/df9b5cd6-8624-48fd-996d-f5d1410c325e)
after running the example script "batch_process_animals.py" it produces the following image among others. There is density but it is very faint and looks quite different than the data shown in the readme.

Can you confirm this is the expected output and if so consider adjusting the colour intensity to make it more visible. I suspect something has gone wrong since in the top right plot it shows expression outside of the brain.
