### 2025-05-22 TeamNCMC/histoquant#32 @PolarBean
readme images aren't viewable when github is in darkmode.
https://github.com/openjournals/joss-reviews/issues/7906
![Image](https://github.com/user-attachments/assets/1f370b0b-b142-4df5-b9f0-6bb3661faac2)
the images in the readme have a transparent background which makes them difficult to view in darkmode. Consider exporting them with a white background.
