### 2025-05-22 TeamNCMC/cuisto#33 @PolarBean
segment images is not self contained.
https://github.com/openjournals/joss-reviews/issues/7906
related to https://github.com/TeamNCMC/cuisto/issues/17

thank you for resolving the reproducability issue with the other example. However it would be nice to expand this to the segment images script.

the demo notebooks are great and are self contained when running from the cloned git repository.
