### 2025-05-22 TeamNCMC/histoquant#34 @PolarBean
consider duplicating demo_notebooks as scripts in the examples folder
https://github.com/openjournals/joss-reviews/issues/7906
at present there are only two examples, though you have many more great self contained examples in the docs/demo_notebooks folder. I understand this is so you can display them in the documentation which is fine. I do think you could help users by including them in the examples folder, potentially reformatted as .py files.
