### 2025-05-26 eb93e00
self-runnable examples as scripts (from existing notebooks)
