### 2025-05-26 c026e00
wip self-running segmentation example
