### 2025-05-26 b50d164
Update segment_images.py, example-seg.tar.gz, and animalid0_001_Probabilities.ome.tiff
