### 2025-05-26 TeamNCMC/cuisto#30 @GuillaumeLeGoc
Wow, I didn't pay attention, that does seem a bit bloated. I don't know how to easily check where dependencies come from. I suspect metapackage such as `notebook`, that is not required per se, but I did want to the user uncomfortable with the command line and Python environments to have everything they could need in one go (including Pandas in performance mode, Brainrender that ships vedo, and so on).
