### 2025-05-27 TeamNCMC/histoquant#39 @PolarBean
new scripts do not run
https://github.com/openjournals/joss-reviews/issues/7906
when running the cells distribution script from both the examples directory and the cuisto directory I'm met with the following error.
```
❯ python cells_distributions.py
Traceback (most recent call last):
  File "/home/harryc/github/cuitso-reviewin/cuisto/examples/cells_distributions.py", line 44, in <module>
    cfg = cuisto.config.Config(config_file)
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/harryc/anaconda3/envs/cuisto-env/lib/python3.12/site-packages/cuisto/config.py", line 59, in __init__
    self.get_blacklist()
  File "/home/harryc/anaconda3/envs/cuisto-env/lib/python3.12/site-packages/cuisto/config.py", line 65, in get_blacklist
    self.atlas["blacklist"] = utils.get_blacklist(
                              ^^^^^^^^^^^^^^^^^^^^
  File "/home/harryc/anaconda3/envs/cuisto-env/lib/python3.12/site-packages/cuisto/utils.py", line 132, in get_blacklist
    with open(file, "rb") as fid:
         ^^^^^^^^^^^^^^^^
FileNotFoundError: [Errno 2] No such file or directory: '../../resources/demo_atlas_blacklist_brain.toml'
```
I assume this is a holdover from when this script was in the docs/demo_notebooks directory. Please test all the scripts and ensure that they run.
