### 2025-05-28 TeamNCMC/cuisto#39 @GuillaumeLeGoc
Hi, this should work now. Would you kindly try again ? You might need to use the local version of the package (`pip install .`) as I needed to tweak the Config class so that it does not crash when the blacklist file is missing (so that we can supply it after instantiating it). I'll make a release if it works for you. Cheers !
