### 2025-05-28 TeamNCMC/cuisto#39 @PolarBean
fantastic, all examples now run as expected.
