### 2025-05-28 TeamNCMC/cuisto#39 @PolarBean
The only issue I have is that segment_images doesnt produce any graphs or say where it's saving data. and when you find the data in the .cuisto folder it just appears black when opening in a standard image viewer. Perhaps make this script communicate something to the user, either through print statements or through having a plot appear which includes whatever it is this script is producing. All other scripts are excellent.
