### 2025-05-28 #7906 @PolarBean
@adamltyson I've finished reviewing the code. Once the manuscript is updated with the changes we discussed I'm happy for this to be published.
