### 2025-05-28 #7906 @IgorTatarnikov
Hi @adamltyson, I've also finished reviewing the code following the requested changes. I'm happy with the updates and I can recommend acceptance.
