### 2025-06-13 77a7684
Update paper.bib
