+++ paper.bib
@@ -240,3 +240,50 @@
   doi = {10.1109/TMI.2009.2035616},
   langid = {english}
 }
+
+@incollection{schmidt2018,
+  title = {Cell {{Detection}} with {{Star-Convex Polygons}}},
+  booktitle = {Medical {{Image Computing}} and {{Computer Assisted Intervention}} – {{MICCAI}} 2018},
+  author = {Schmidt, Uwe and Weigert, Martin and Broaddus, Coleman and Myers, Gene},
+  editor = {Frangi, Alejandro F. and Schnabel, Julia A. and Davatzikos, Christos and Alberola-López, Carlos and Fichtinger, Gabor},
+  date = {2018},
+  volume = {11071},
+  pages = {265--273},
+  publisher = {Springer International Publishing},
+  location = {Cham},
+  doi = {10.1007/978-3-030-00934-2_30},
+  url = {https://link.springer.com/10.1007/978-3-030-00934-2_30},
+  urldate = {2023-09-14},
+  abstract = {Automatic detection and segmentation of cells and nuclei in microscopy images is important for many biological applications. Recent successful learning-based approaches include per-pixel cell segmentation with subsequent pixel grouping, or localization of bounding boxes with subsequent shape refinement. In situations of crowded cells, these can be prone to segmentation errors, such as falsely merging bordering cells or suppressing valid cell instances due to the poor approximation with bounding boxes. To overcome these issues, we propose to localize cell nuclei via star-convex polygons, which are a much better shape representation as compared to bounding boxes and thus do not need shape refinement. To that end, we train a convolutional neural network that predicts for every pixel a polygon for the cell instance at that position. We demonstrate the merits of our approach on two synthetic datasets and one challenging dataset of diverse fluorescence microscopy images.},
+  isbn = {978-3-030-00933-5 978-3-030-00934-2},
+  langid = {english},
+}
+
+@online{goldsborough2024a,
+  title = {A Novel Channel Invariant Architecture for the Segmentation of Cells and Nuclei in Multiplexed Images Using {{InstanSeg}}},
+  author = {Goldsborough, Thibaut and O’Callaghan, Alan and Inglis, Fiona and Leplat, Léo and Filby, Andrew and Bilen, Hakan and Bankhead, Peter},
+  date = {2024-09-08},
+  doi = {10.1101/2024.09.04.611150},
+  url = {http://biorxiv.org/lookup/doi/10.1101/2024.09.04.611150},
+  urldate = {2024-09-12},
+  abstract = {The quantitative analysis of bioimaging data increasingly depends on the accurate segmentation of cells and nuclei, a significant challenge for the analysis of high-plex imaging data. Current deep learning-based approaches to segment cells in multiplexed images require reducing the input to a small and fixed number of input channels, discarding imaging information in the process. We present ChannelNet, a novel deep learning architecture for generating threechannel representations of multiplexed images irrespective of the number or ordering of imaged biomarkers. When combined with InstanSeg, ChannelNet sets a new benchmark for the segmentation of cells and nuclei on public multiplexed imaging datasets. We provide an open implementation of our method and integrate it in open source software. Our code and models are available on https://github.com/instanseg/instanseg.},
+  langid = {english},
+  pubstate = {prepublished},
+}
+
+@article{stringer2021,
+  title = {Cellpose: A Generalist Algorithm for Cellular Segmentation},
+  shorttitle = {Cellpose},
+  author = {Stringer, Carsen and Wang, Tim and Michaelos, Michalis and Pachitariu, Marius},
+  date = {2021-01},
+  journaltitle = {Nature Methods},
[…]
