+++ paper.md
@@ -33,19 +33,19 @@ bibliography: paper.bib
 # Summary
 Fluorescent labeling techniques, including immunohistochemistry and endogenously fluorescent proteins, are key assets in neuroscience. Combined to genetic toolboxes, they enable the anatomical identification of specific neurons and neuronal processes, shedding light on neural networks organization and linking behavior to brain structures. Advances in imaging techniques and numerical tools have enabled the creation of volumetric, annotated, whole-brain atlases of various animal models [@wang2020; @kleven2023; @kunst2019; @lazcano2021] facilitating brain-wide mapping of labeled elements to a reference three-dimensional space. In this framework, a common task is to count, in the reference brain regions, objects of interest (be it whole cells, nuclei, axons, synaptic puncta...) detected in 2D histological slices imaged with fluorescence microscopy.
 
-Multiple software solutions exist to perform image analysis for objects detection, and tools to register 2D slices to 3D reference atlases are also available. The `cuisto` package aims at bridging these pieces of software together to provide a streamlined process, from raw images to quantification figures. It harnesses the output of the bioimage analysis software QuPath [@bankhead2017] used together with the registration toolbox ABBA [@chiaruttini2024]. Designed with users who have minimal programming knowledge in mind, it is configurable and intended to be atlas- and staining-agnostic to cover various use-cases, and provides utility scripts that can be used throughout the analysis pipeline. Furthermore, an extensive documentation is provided, and includes 1) the installation of the various pieces of software used upstream of the `cuisto` package, 2) ABBA and QuPath tutorials, 3) in-depth explanation of the data formatting requirements and 4) hands-on examples with Jupyter notebooks.
+Multiple software solutions exist to perform image analysis for objects detection, and tools to register 2D slices to 3D reference atlases are also available. The `cuisto` package aims at bridging these pieces of software together to provide a streamlined process, from raw images to quantification figures. It harnesses the output of the bioimage analysis software QuPath [@bankhead2017] used together with the registration toolbox ABBA [@chiaruttini2024]. Designed with users who have minimal programming knowledge in mind, it is configurable and intended to be atlas- and staining-agnostic to cover various use-cases, and provides utility scripts that can be used throughout the analysis pipeline. Furthermore, an extensive documentation is provided, and includes 1) the installation of the various pieces of software used upstream of the `cuisto` package, 2) ABBA and QuPath tutorials, 3) in-depth explanation of the data formatting requirements and 4) hands-on examples with Python scripts and Jupyter notebooks.
 
 # Statement of need
-As the task of counting objects in brain regions from fluorescent microscopy images is widespread, several toolboxes were specifically designed for brain-wide mapping. Regrettably however, these toolboxes often lack maintenance and/or community support, as is the case with the QUINT workflow [@yates2019] that relies on several pieces of software (including QuickNII [@puchades2019] and Ilastik [@berg2019]). Furthermore, while it was designed precisely for counting fluorescent objects, it can be rigid at times, making hard to take shortcuts during the workflow and is not interoperable with other computational neuroanatomical tools, notably reference atlases provided by Brainglobe [@claudi2020]. The latter does provide numerous tools to perform brain-wide cell counting [@tyson2021; @tyson2022] but is primarily designed for native 3D data (such as obtained from Light Sheet imaging post clearing). This approach is however not readily accessible to all laboratories and may show variable efficacy depending on the fluorophore used,  the intended resolution or the targeted brain structures.
[…]
