+++ paper.md
@@ -38,7 +38,7 @@ Multiple software solutions exist to perform image analysis for objects detectio
 # Statement of need
 As the task of counting objects in brain regions from fluorescent microscopy images is widespread, several toolboxes have been developed for brain-wide object quantification such as the QUINT workflow [@yates2019]. The pipeline relies on QuickNII and VisuAlign [@puchades2019] for the registration and Ilastik [berg2019] for the segmentation. While effective for specific tasks, such as counting punctal objects, its architecture offers limited flexibility, making it hard to take shortcuts during the workflow, and it is not interoperable with other computational neuroanatomical tools, notably reference atlases provided by Brainglobe [@claudi2020]. The latter does provide numerous tools to perform brain-wide cell counting [@tyson2021; @tyson2022] but is primarily designed for native 3D data (such as obtained from Light Sheet imaging post clearing). This approach is however not readily accessible to all laboratories and may show variable efficacy depending on the fluorophore used, the intended resolution, or the targeted brain structures.
 
-The QuPath software provides a full-featured, user-friendly interface to perform image quantification while being extensible and scriptable. It is being actively developed and is supported by a vivid community (more than 4k topics tagged with "qupath" on the [image.sc forum](https://forum.image.sc/tag/qupath) and more than 4k citations). This software  supports a variety of segmentation strategies, from basic thresholding to pixel classification and advanced deep learning methods such as CellPose [@stringer2021], StarDist [@schmidt2018],and InstanSeg [@goldsborough2024a]. The Fiji [@schindelin2012] plugin Aligning Big Brain and Atlases (ABBA) allows for semi-automated registration of whole-brain 2D sections to a 3D atlas in an intuitive and interactive graphical user interface using native full-resolution multichannel images, providing both automatic in-plane registration through elastix [@klein2010] and manual adjustment with BigWarp [@bogovic2016]. It also supports the deep-learning-based automatic registration tool DeepSlice [@carey2023] and can be interfaced with Brainglobe atlases. Furthermore, it integrates seamlessly with QuPath, to both import images and export back the registration results as annotations in QuPath for further quantification.
+The QuPath software provides a full-featured, user-friendly interface to perform image quantification while being extensible and scriptable. It is being actively developed and is supported by a vivid community (more than 4k topics tagged with "qupath" on the [image.sc forum](https://forum.image.sc/tag/qupath) and more than 4k citations). This software  supports a variety of segmentation strategies, from basic thresholding to pixel classification and advanced deep learning methods such as CellPose [@stringer2021], StarDist [@schmidt2018] and InstanSeg [@goldsborough2024a]. The Fiji [@schindelin2012] plugin Aligning Big Brain and Atlases (ABBA) allows for semi-automated registration of whole-brain 2D sections to a 3D atlas in an intuitive and interactive graphical user interface using native full-resolution multichannel images, providing both automatic in-plane registration through elastix [@klein2010] and manual adjustment with BigWarp [@bogovic2016]. It also supports the deep-learning-based automatic registration tool DeepSlice [@carey2023] and can be interfaced with Brainglobe atlases. Furthermore, it integrates seamlessly with QuPath, to both import images and export back the registration results as annotations in QuPath for further quantification.
 
 Yet, to our knowledge, no streamlined pipeline existed to bridge the image processing and registration with the region-based quantification and data-visualization. The present work is two-fold :
 
