### 2025-06-13 #7906 @GuillaumeLeGoc
Hi @adamltyson, @PolarBean, @IgorTatarnikov,
Again, sorry for the delay. We updated the manuscript to address your inputs, including :
- Removed false statement about QUINT maintenance status
- Expanded a bit on QuPath features to illustrate why it's a good alternative
- Some mentions about segmentation and brain contours were added to reflect the recent additions

I wish you a good weekend !
