### 2025-06-13 #7906 @PolarBean
I think this is improved. Thanks for your work on this. I'm happy for it to be published as is.
