### 2025-06-17 7f2d6a8
Update paper.md
