+++ paper.md
@@ -36,18 +36,18 @@ Fluorescent labeling techniques, including immunohistochemistry and endogenously
 Multiple software solutions exist to perform image analysis for objects detection, and tools to register 2D slices to 3D reference atlases are also available. The `cuisto` package aims at bridging these pieces of software together to provide a streamlined process, from raw images to quantification figures. It harnesses the output of the bioimage analysis software QuPath [@bankhead2017] used together with the registration toolbox ABBA [@chiaruttini2024]. Designed with users who have minimal programming knowledge in mind, it is configurable and intended to be atlas- and staining-agnostic to cover various use-cases, and provides utility scripts that can be used throughout the analysis pipeline. Furthermore, an extensive documentation is provided, and includes 1) the installation of the various pieces of software used upstream of the `cuisto` package, 2) ABBA and QuPath tutorials, 3) in-depth explanation of the data formatting requirements and 4) hands-on examples with Python scripts and Jupyter notebooks.
 
 # Statement of need
-As the task of counting objects in brain regions from fluorescent microscopy images is widespread, several toolboxes have been developed for brain-wide object quantification such as the QUINT workflow [@yates2019]. The pipeline relies on QuickNII and VisuAlign [@puchades2019] for the registration and Ilastik [@berg2019] for the segmentation. While effective for specific tasks, such as counting punctal objects, its architecture offers limited flexibility, making it hard to take shortcuts during the workflow, and it is not interoperable with other computational neuroanatomical tools, notably reference atlases provided by Brainglobe [@claudi2020]. The latter does provide numerous tools to perform brain-wide cell counting [@tyson2021; @tyson2022] but is primarily designed for native 3D data (such as obtained from Light Sheet imaging post clearing). This approach is however not readily accessible to all laboratories and may show variable efficacy depending on the fluorophore used, the intended resolution, or the targeted brain structures.
+As the task of counting objects in brain regions from fluorescent microscopy images is widespread, several toolboxes have been developed for brain-wide object quantification such as the QUINT workflow [@yates2019]. The pipeline relies on QuickNII and VisuAlign [@puchades2019] for the registration and Ilastik [@berg2019] for the segmentation. While effective for specific tasks, such as counting punctal objects, its architecture offers limited flexibility, making it hard to take shortcuts during the workflow, and it is not interoperable with other computational neuroanatomical tools, notably reference atlases provided by BrainGlobe [@claudi2020]. The latter does provide numerous tools to perform brain-wide cell counting [@tyson2021; @tyson2022] but is primarily designed for native 3D data (such as obtained from light sheet imaging of opticaly cleared tissue). This approach is however not readily accessible to all laboratories and may show variable efficacy depending on the fluorophore used, the intended resolution, or the targeted brain structures.
 
[…]
