### 2025-06-17 #7906 @adamltyson
Hi @GuillaumeLeGoc, looks like we're nearly there. I've raised https://github.com/TeamNCMC/cuisto/pull/42 with some paper suggestions. Once you've accepted or rejected that PR, could you complete the "Additional Author Tasks After Review is Complete" listed above?
