### 2025-06-17 #7906 @adamltyson
Your release doesn't need to include the paper itself (that will end up on the JOSS website). You can release from main (or wherever you normally release from), as long as all changes from this review process are included in that branch.
