+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "cuisto"
-version = "2025.5.28"
+version = "2025.6.17"
 authors = [{ name = "Guillaume Le Goc", email = "g.legoc@posteo.org" }]
 description = "Quantification of objects in histological slices"
 readme = "README.md"
