### 2025-06-26 #7906 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [ ] check that all needed metadata is present for this bioRxiv paper in the references (formatting looks odd to me): https://www.biorxiv.org/content/10.1101/2024.09.04.611150v1
- [ ] first line of second paragraph of "Summary" section: should be "object detection"
- [ ] later in that paragraph: remove "an" so sentence begins "Furthermore, extensive documentation is provided..."
- [ ] third line of first full paragraph on second page: suggest replacing "vivid" with "vibrant"
- [ ] next paragraph: remove hyphen from "data visualization"
- [ ] paragraph below numbered list: add missing Oxford comma after "polygons"
- [ ] same paragraph: replace "cumulated" with "cumulative" (also in next paragraph) unless "cumulated" is very standard usage in this field
- [ ] same paragraph: suggest replacing "fibers-like" with "fiber-like" in both instances unless "fibers-like" is very standard usage in this field
- [ ] following paragraph: replace "atlas regions names" with "atlas region names" and in paragraph below bullet list, same thing but with "contours"
- [ ] same paragraph: another missing Oxford comma after "names"
- [ ] bullet list below that: add "and" after second bullet point and replace both instances of "e.g." with "i.e."
- [ ] Acknowledgements: add comma after "Bunkhead"
- [ ] replace "feedbacks" with "feedback"
