### 2025-06-27 a9b0b91
fixes from joss aeic
