+++ paper.bib
@@ -1,13 +1,14 @@
-@online{chiaruttini2024,
-  title = {{{ABBA}}, a Novel Tool for Whole-Brain Mapping, Reveals Brain-Wide Differences in Immediate Early Genes Induction Following Learning},
-  author = {Chiaruttini, Nicolas and Castoldi, Carlo and Requie, Linda Maria and Camarena-Delgado, Carmen and Dal Bianco, Beatrice and Gräff, Johannes and Seitz, Arne and Silva, Bianca A.},
-  date = {2024-09-06},
-  doi = {10.1101/2024.09.06.611625},
-  url = {http://biorxiv.org/lookup/doi/10.1101/2024.09.06.611625},
-  urldate = {2024-09-12},
-  abstract = {Abstract           Unbiased characterization of whole-brain cytoarchitecture represents an invaluable tool for understanding brain function. For this, precise mapping of histological markers from 2D sections onto 3D brain atlases is pivotal. Here, we present two novel software tools facilitating this process: Aligning Big Brains and Atlases (ABBA), designed to streamline the precise and efficient registration of 2D sections to 3D reference atlases, and BraiAn, an integrated suite for multi-marker automated segmentation, whole-brain statistical analysis, and data visualisation. Combining these tools, we performed a comprehensive comparative study of the whole-brain expression of three of the most widely used immediate early genes (IEGs). Thanks to their neural activity-dependent expression, IEGs have been used for decades as a proxy of neural activity to generate unbiased mapping of activity following behaviour, but their respective induction in response to neuronal activation across the entire brain remains unclear. To address this question, we systematically compared the brain-wide expression cFos, Arc and NPAS4, three abundantly used IEGs, across three different behavioural conditions related to memory. Our results highlight major differences in both their distribution and induction patterns, indicating that they do not represent equivalent markers across brain areas or activity states, but can provide instead complementary information.},
-  langid = {english},
-  pubstate = {prepublished}
+@article {chiaruttini2024,
+	author = {Chiaruttini, Nicolas and Castoldi, Carlo and Requie, Linda Maria and Camarena-Delgado, Carmen and dal Bianco, Beatrice and Gr{\"a}ff, Johannes and Seitz, Arne and Silva, Bianca A.},
+	title = {ABBA, a novel tool for whole-brain mapping, reveals brain-wide differences in immediate early genes induction following learning},
+	elocation-id = {2024.09.06.611625},
+	year = {2024},
+	doi = {10.1101/2024.09.06.611625},
+	publisher = {Cold Spring Harbor Laboratory},
[…]
