### 2025-06-27 #7906 @GuillaumeLeGoc
Hi @rkurchin, thanks for taking over this submission !
Thank you for your suggestions, they should be fixed now. I went ahead and fix the bibtex for Chiaruttini 2024 that was odd as well (they were set as online resources, they are now articles).
Please let me know if I overlooked anything. Cheers !
