+++ paper.bib
@@ -1,16 +1,18 @@
-@article {chiaruttini2024,
-	author = {Chiaruttini, Nicolas and Castoldi, Carlo and Requie, Linda Maria and Camarena-Delgado, Carmen and dal Bianco, Beatrice and Gr{\"a}ff, Johannes and Seitz, Arne and Silva, Bianca A.},
-	title = {ABBA, a novel tool for whole-brain mapping, reveals brain-wide differences in immediate early genes induction following learning},
-	elocation-id = {2024.09.06.611625},
-	year = {2024},
-	doi = {10.1101/2024.09.06.611625},
-	publisher = {Cold Spring Harbor Laboratory},
-	abstract = {Unbiased characterization of whole-brain cytoarchitecture represents an invaluable tool for understanding brain function. For this, precise mapping of histological markers from 2D sections onto 3D brain atlases is pivotal. Here, we present two novel software tools facilitating this process: Aligning Big Brains and Atlases (ABBA), designed to streamline the precise and efficient registration of 2D sections to 3D reference atlases, and BraiAn, an integrated suite for multi-marker automated segmentation, whole-brain statistical analysis, and data visualisation. Combining these tools, we performed a comprehensive comparative study of the whole-brain expression of three of the most widely used immediate early genes (IEGs). Thanks to their neural activity-dependent expression, IEGs have been used for decades as a proxy of neural activity to generate unbiased mapping of activity following behaviour, but their respective induction in response to neuronal activation across the entire brain remains unclear. To address this question, we systematically compared the brain-wide expression cFos, Arc and NPAS4, three abundantly used IEGs, across three different behavioural conditions related to memory. Our results highlight major differences in both their distribution and induction patterns, indicating that they do not represent equivalent markers across brain areas or activity states, but can provide instead complementary information.Competing Interest StatementThe authors have declared no competing interest.},
-	URL = {https://www.biorxiv.org/content/early/2024/09/06/2024.09.06.611625},
-	eprint = {https://www.biorxiv.org/content/early/2024/09/06/2024.09.06.611625.full.pdf},
-	journal = {bioRxiv}
+@article{chiaruttini2025,
+	title = {{ABBA}+{BraiAn}, an integrated suite for whole-brain mapping, reveals brain-wide differences in immediate-early genes induction upon learning},
+	volume = {44},
+	issn = {22111247},
+	url = {https://linkinghub.elsevier.com/retrieve/pii/S2211124725006473},
+	doi = {10.1016/j.celrep.2025.115876},
+	language = {en},
+	number = {7},
+	urldate = {2025-06-27},
+	journal = {Cell Reports},
+	author = {Chiaruttini, Nicolas and Castoldi, Carlo and Requie, Linda Maria and Camarena-Delgado, Carmen and Dal Bianco, Beatrice and Gräff, Johannes and Seitz, Arne and Silva, Bianca A.},
+	month = jul,
+	year = {2025},
+	pages = {115876},
 }
-
 @article{bankhead2017,
   title = {{{QuPath}}: {{Open}} Source Software for Digital Pathology Image Analysis},
   shorttitle = {{{QuPath}}},
