### 2025-06-27 #7906 @GuillaumeLeGoc
Nice catch @PolarBean, thanks, the reference has been updated accordingly.
