### 2025-07-03 #7906 @GuillaumeLeGoc
Hi @rkurchin, it should be fixed. Cheers !
