### 2025-07-04 c311c65
Create CITATION.cff
