### 2025-07-04 25ef1e7
Added JOSS badge
